bugfix: lseek should return off_t
lseek was triggering a number overflow by converting long to integer on
its return value

Added specific lseek test

+++ include/client/hooks.hpp
@@ -34,7 +34,7 @@ int hook_unlinkat(int dirfd, const char * cpath, int flags);
 int hook_symlinkat(const char * oldname, int newdfd, const char * newname);
 int hook_access(const char* path, int mask);
 int hook_faccessat(int dirfd, const char * cpath, int mode);
-int hook_lseek(unsigned int fd, off_t offset, unsigned int whence);
+off_t hook_lseek(unsigned int fd, off_t offset, unsigned int whence);
 int hook_truncate(const char *path, long length);
 int hook_ftruncate(unsigned int fd, unsigned long length);
 int hook_dup(unsigned int fd);
